[Paddle Tensorrt] change int32 to int64 in test file (#69191)
* mod int64

* rm pass_test

* change np

* change file

* fix bug

* fix conflict

* add test mod

* Update

* fix conflict

* update third_party

* update flashattn

* change type

+++ test/tensorrt/test_converter_common.py
@@ -225,8 +225,8 @@ class TestNearestSizeTensorTRTPattern(TensorRTBaseTest):
             "x": x_nchw,
             "OutSize": None,
             "SizeTensor": [
-                np.array([12], dtype="int32"),
-                np.array([12], dtype="int32"),
+                np.array([12], dtype="int64"),
+                np.array([12], dtype="int64"),
             ],
             "Scale": None,
             "attrs": {
